rename staff list from index
